atualizando

+++ src/controllers/HouseController.js
@@ -17,7 +17,7 @@ class HouseController{
 
     const house = await House.create({
       user: user_id,
-      description,
+            description,
       price,
       location,
       status,
@@ -27,18 +27,15 @@ class HouseController{
   }
 
   async update(req, res){
-      const { house_id } = req.params;
+    const { house_id } = req.params;
     const { description, price, location, status } = req.body;
     const { user_id } = req.headers;
+
 
     const user = await User.findById(user_id);
     const houses = await House.findById(house_id);
 
-    if(String(user._id) !== String(houses.user)){
-      return res.status(401).json({ error: 'Não autorizado.'});
-    }
-
-    await House.updateOne({ _id: house_id }, {
+    await House.updateOne({ _id: house_id  }, {
       user: user_id,
       description,
       price,
@@ -52,11 +49,21 @@ class HouseController{
 
   async destroy(req, res){
     const { house_id } = req.body;
-   
-    await House.findByIdAndDelete({_id: house_id});
+    const { user_id } = req.headers;
+
+    const user = await User.findById(user_id);
+    const houses = await House.findById(house_id);
+
+    if(String(user._id) !== String(houses.user)){
+      return res.status(401).json({ error: 'Não autorizado.'});
+    }
+
+    await House.findByIdAndDelete({ _id: house_id});
 
     return res.json({message: "Excluida com sucesso!" });
   }
+
 }
-    export default new HouseController();
+
+export default new HouseController();
